+++ src/i18n/en.extra.js
@@ -2,5 +2,8 @@ window.ParsleyConfig = window.ParsleyConfig || {};
 window.ParsleyConfig.i18n = window.ParsleyConfig.i18n || {};
 
 window.ParsleyConfig.i18n.en = $.extend(window.ParsleyConfig.i18n.en || {}, {
-  dateiso: "This value should be a valid date (YYYY-MM-DD)."
+  dateiso:  "This value should be a valid date (YYYY-MM-DD).",
+  minwords: "This value is too short. It should have %s words or more.",
+  maxwords: "This value is too long. It should have %s words or fewer.",
+  words:    "This value length is invalid. It should be between %s and %s words long."
 });